+++ containers/fhm.h
@@ -2,6 +2,8 @@
 // open horizon -- undefined_darkness@outlook.com
 //
 
+// internal container since ac6
+
 #pragma once
 
 #include "resources/resources.h"
@@ -20,11 +22,10 @@ public:
 
     int get_chunks_count() const { return int(m_chunks.size()); }
 
-    unsigned int get_chunk_type(int idx) const;
-    size_t get_chunk_size(int idx) const;
+    uint32_t get_chunk_type(int idx) const;
+    uint32_t get_chunk_size(int idx) const;
     bool read_chunk_data(int idx, void *data) const;
-
-    size_t get_chunk_offset(int idx) const;
+    uint32_t get_chunk_offset(int idx) const;
 
     fhm_file(): m_data(0) {}
 
@@ -54,9 +55,9 @@ private:
 
     struct chunk
     {
-        unsigned int type;
-        size_t offset;
-        size_t size;
+        uint32_t type;
+        uint32_t offset;
+        uint32_t size;
     };
 
     std::vector<chunk> m_chunks;
